Move app-config to use service TraceProvider.
This PR moves the app-config service to use the service TraceProvider.
