+++ services/app-provider/pkg/command/server.go
@@ -12,13 +12,13 @@ import (
 	"github.com/owncloud/ocis/v2/ocis-pkg/config/configlog"
 	"github.com/owncloud/ocis/v2/ocis-pkg/registry"
 	"github.com/owncloud/ocis/v2/ocis-pkg/sync"
+	"github.com/owncloud/ocis/v2/ocis-pkg/tracing"
 	"github.com/owncloud/ocis/v2/ocis-pkg/version"
 	"github.com/owncloud/ocis/v2/services/app-provider/pkg/config"
 	"github.com/owncloud/ocis/v2/services/app-provider/pkg/config/parser"
 	"github.com/owncloud/ocis/v2/services/app-provider/pkg/logging"
 	"github.com/owncloud/ocis/v2/services/app-provider/pkg/revaconfig"
 	"github.com/owncloud/ocis/v2/services/app-provider/pkg/server/debug"
-	"github.com/owncloud/ocis/v2/services/app-provider/pkg/tracing"
 	"github.com/urfave/cli/v2"
 )
 
@@ -33,7 +33,7 @@ func Server(cfg *config.Config) *cli.Command {
 		},
 		Action: func(c *cli.Context) error {
 			logger := logging.Configure(cfg.Service.Name, cfg.Log)
-			err := tracing.Configure(cfg, logger)
+			tracingProvider, err := tracing.GetServiceTraceProvider(cfg.Tracing, cfg.Service.Name)
 			if err != nil {
 				return err
 			}
@@ -50,6 +50,7 @@ func Server(cfg *config.Config) *cli.Command {
 				runtime.RunWithOptions(rCfg, pidFile,
 					runtime.WithLogger(&logger.Logger),
 					runtime.WithRegistry(reg),
+					runtime.WithTraceProvider(tracingProvider),
 				)
 
 				return nil
@@ -67,7 +68,6 @@ func Server(cfg *config.Config) *cli.Command {
 				debug.Context(ctx),
 				debug.Config(cfg),
 			)
-
 			if err != nil {
 				logger.Info().Err(err).Str("server", "debug").Msg("Failed to initialize server")
 				return err
