+++ services/app-provider/pkg/revaconfig/config.go
@@ -6,15 +6,7 @@ import (
 
 // AppProviderConfigFromStruct will adapt an oCIS config struct into a reva mapstructure to start a reva service.
 func AppProviderConfigFromStruct(cfg *config.Config) map[string]interface{} {
-
 	rcfg := map[string]interface{}{
-		"core": map[string]interface{}{
-			"tracing_enabled":      cfg.Tracing.Enabled,
-			"tracing_exporter":     cfg.Tracing.Type,
-			"tracing_endpoint":     cfg.Tracing.Endpoint,
-			"tracing_collector":    cfg.Tracing.Collector,
-			"tracing_service_name": cfg.Service.Name,
-		},
 		"shared": map[string]interface{}{
 			"jwt_secret":          cfg.TokenManager.JWTSecret,
 			"gatewaysvc":          cfg.Reva.Address,
